requirements.txt
djangorestframework==3.15.1
drf-yasg==1.21.7
django==4.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn
